Simplify package-template usage and updating by requiring fewer files to be edited.
