+++ packagename/tests/coveragerc
@@ -1,5 +1,5 @@
 [run]
-source = packagename
+source = {packagename}
 
 [report]
 exclude_lines =
