+++ setup.cfg
@@ -10,3 +10,14 @@ show-response = 1
 [pytest]
 minversion = 2.2
 norecursedirs = build docs/_build
+
+[affiliated]
+package_name = packagename
+description = Astropy affiliated package
+long_description = This is a package for doing X that relies on astropy
+author = John Doe
+author_email = johndoe@nowhere.com
+license = BSD
+url = http://astropy.org/
+edit_on_github = False
+github_project = astropy/astropy
