+++ setup.py
@@ -20,14 +20,27 @@ from astropy.setup_helpers import (register_commands, adjust_compiler,
                                    get_debug_option)
 from astropy.version_helpers import get_git_devstr, generate_version_py
 
-# Set affiliated package-specific settings
-PACKAGENAME = 'packagename'
-DESCRIPTION = 'Astropy affiliated package'
-LONG_DESCRIPTION = ''
-AUTHOR = ''
-AUTHOR_EMAIL = ''
-LICENSE = 'BSD'
-URL = 'http://astropy.org'
+# Get some values from the setup.cfg
+from distutils import config
+conf = config.ConfigParser()
+conf.read(['setup.cfg'])
+affiliated = dict(conf.items('affiliated'))
+
+PACKAGENAME = affiliated.get('package_name', 'packagename')
+DESCRIPTION = affiliated.get('description', 'Astropy affiliated package')
+AUTHOR = affiliated.get('author', '')
+AUTHOR_EMAIL = affiliated.get('author_email', '')
+LICENSE = affiliated.get('license', 'unknown')
+URL = affiliated.get('url', 'http://astropy.org')
+
+# Get the long description from the package's docstring
+__import__(PACKAGENAME)
+package = sys.modules[PACKAGENAME]
+LONG_DESCRIPTION = package.__doc__
+
+# Store the package name in a built-in variable so it's easy
+# to get from other parts of the setup infrastructure
+builtins._ASTROPY_PACKAGE_NAME_ = PACKAGENAME
 
 # VERSION should be PEP386 compatible (http://www.python.org/dev/peps/pep-0386)
 VERSION = '0.0.dev'
@@ -94,6 +107,16 @@ except ImportError: # compatibility with Astropy 0.2 - can be removed in cases
     update_package_files(PACKAGENAME, package_info['ext_modules'],
                          package_info['package_data'], package_info['packages'],
                          package_info['package_dir'])
+
+# Include all .c files, recursively
+c_files = []
+for root, dirs, files in os.walk(PACKAGENAME):
+    for filename in files:
+        if filename.endswith('.c'):
+            c_files.append(
+                os.path.join(
+                    os.path.relpath(root, PACKAGENAME), filename))
+package_info['package_data'][PACKAGENAME].extend(c_files)
 
 setup(name=PACKAGENAME,
       version=VERSION,
